requirements.txt
compas_assembly >=0.7.0
compas_cra
compas_dr
compas_fd
compas_tna
compas_ifc
compas_model
compas_viewer
compas_notebook
